changed blockquote to italic
i have been back and forth on this. I have a desire to limit the number of categories implied by text styling options.
dark/light is markup vs content
saturated/greyscale is links and tags vs everything else
I have read that italics are difficult for people with dyslexia to read, but at the same time it is a commonly understood option for communicating certain types of info. My thought is if wide letter spacing is used along with a font with a true (not oblique) italic it will minimize difficulty